server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
config==0.5.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
filelock==3.13.1
Flask==3.1.0
flask-cors==5.0.1
Flask-PyMongo==3.0.1
Flask-SocketIO==5.5.1
fonttools==4.57.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
openai==1.61.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclean==3.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==70.2.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
